+++ src/main/java/com/lovable/lovable/dtos/project/ProjectResponse.java
@@ -2,12 +2,11 @@ package com.lovable.lovable.dtos.project;
 
 import com.lovable.lovable.dtos.auth.UserProfileResponse;
 
-import java.math.BigInteger;
 import java.time.Instant;
 
 public record ProjectResponse(
 
-        BigInteger id,
+        Long id,
         String name,
         Instant createdAt,
         Instant updatedAt,
