+++ src/main/java/com/lovable/lovable/service/FileService.java
@@ -3,11 +3,11 @@ package com.lovable.lovable.service;
 import com.lovable.lovable.dtos.project.FileContentResponse;
 import com.lovable.lovable.dtos.project.FileNode;
 
-import java.math.BigInteger;
+
 import java.util.List;
 
 public interface FileService {
-    List<FileNode> getFileTree(BigInteger userId, BigInteger projectId);
+    List<FileNode> getFileTree(Long userId, Long projectId);
 
-    FileContentResponse getFileContent(BigInteger userId, BigInteger projectId, String path);
+    FileContentResponse getFileContent(Long userId, Long projectId, String path);
 }
