+++ src/main/java/com/lovable/lovable/service/UsageService.java
@@ -3,10 +3,9 @@ package com.lovable.lovable.service;
 import com.lovable.lovable.dtos.subscription.PlanLimitResponse;
 import com.lovable.lovable.dtos.subscription.UsageTodayResponse;
 
-import java.math.BigInteger;
 
 public interface UsageService {
-    UsageTodayResponse getTodaysUsage(BigInteger userId);
+    UsageTodayResponse getTodaysUsage(Long userId);
 
-    PlanLimitResponse getCurrentSubscriptionLimits(BigInteger userId);
+    PlanLimitResponse getCurrentSubscriptionLimits(Long userId);
 }
