+++ src/main/java/com/lovable/lovable/service/UserService.java
@@ -2,8 +2,7 @@ package com.lovable.lovable.service;
 
 import com.lovable.lovable.dtos.auth.UserProfileResponse;
 
-import java.math.BigInteger;
 
 public interface UserService {
-    UserProfileResponse getProfile(BigInteger userId);
+    UserProfileResponse getProfile(Long userId);
 }
